evaluation/requirements.txt
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Faker==30.8.1
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
rstr==3.2.2
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3

